+++ src/nespresso/main.py
@@ -1,16 +1,18 @@
 import asyncio
 
 from nespresso.bot.creator import bot, dp
-from nespresso.core import logs
 
-# from db.connect import close_mongo_connection, setup_mongo_connection
 # from handlers.admin import admin
 # from handlers.admin.send_on import send_shutdown, send_startup
 # from handlers.client import client
-# from handlers.client.email import test_emails
 # from handlers.client.menu import set_commands
 # from handlers.common import common_handlers
 # from handlers.common.pending import notify_users_with_pending_updates
+from nespresso.bot.handlers.common.register import (
+    RegisterHandlerCancel,
+    RegisterHandlerZeroMessage,
+)
+from nespresso.core import logs
 
 
 async def OnStartUp() -> None:
@@ -40,10 +42,10 @@ async def main() -> None:
     Registers handlers, sets commands, and starts polling for updates.
     """
     try:
-        # common_handlers.register_handler_cancel(dp)
+        RegisterHandlerZeroMessage(dp)
         # admin.register_handlers_admin(dp)
         # client.register_handlers_client(dp)
-        # common_handlers.register_handler_zero_message(dp)
+        RegisterHandlerCancel(dp)
 
         dp.startup.register(OnStartUp)
         dp.shutdown.register(OnShutdown)