+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Structure of DB models
+- Cancel & ZeroMessage handlers
 
 ## [0.1.0] - 2025-05-07 - Vlad
 ### Added